Apply redux tookit
