+++ package.json
@@ -59,6 +59,7 @@
     "@codeceptjs/configure": "^0.5.2",
     "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
+    "@reduxjs/toolkit": "^1.4.0",
     "facepaint": "^1.2.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
@@ -67,7 +68,6 @@
     "react-router-dom": "^5.2.0",
     "react-slick": "^0.27.4",
     "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
     "slick-carousel": "^1.8.1"
   }
 }
