--- src/actions.test.js
@@ -1,47 +1,0 @@
-import thunk from 'redux-thunk';
-
-import configureStore from 'redux-mock-store';
-
-import {
-  setProducts,
-  setProduct,
-  loadInitProducts,
-  loadProduct,
-} from './actions';
-
-const middlewares = [thunk];
-const mockStore = configureStore(middlewares);
-
-jest.mock('./services/api');
-
-describe('actions', () => {
-  let store;
-
-  describe('loadInitProducts', () => {
-    beforeEach(() => {
-      store = mockStore({});
-    });
-
-    it('runs setProducts', async () => {
-      await store.dispatch(loadInitProducts());
-
-      const actions = store.getActions();
-
-      expect(actions[0]).toEqual(setProducts([]));
-    });
-  });
-
-  describe('loadProduct', () => {
-    beforeEach(() => {
-      store = mockStore({});
-    });
-
-    it('dispatchs setProduct', async () => {
-      await store.dispatch(loadProduct({ productId: 1 }));
-
-      const actions = store.getActions();
-
-      expect(actions[0]).toEqual(setProduct({}));
-    });
-  });
-});
