+++ src/slice.test.js
@@ -1,0 +1,89 @@
+import configureStore from 'redux-mock-store';
+
+import { getDefaultMiddleware } from '@reduxjs/toolkit';
+
+import reducer, {
+  loadInitProducts,
+  loadProduct,
+  setProduct,
+  setProducts,
+} from './slice';
+
+import products from '../fixtures/products';
+
+const middlewares = [...getDefaultMiddleware()];
+const mockStore = configureStore(middlewares);
+
+jest.mock('./services/api');
+
+describe('reducer', () => {
+  context('when previous state is undefined', () => {
+    const initialState = {
+      product: null,
+      products: [],
+    };
+
+    it('returns initialState', () => {
+      const state = reducer(undefined, { type: 'action' });
+
+      expect(state).toEqual(initialState);
+    });
+  });
+
+  describe('setProducts', () => {
+    const initialState = {
+      products: [],
+    };
+
+    it('changes products', () => {
+      const state = reducer(initialState, setProducts(products));
+
+      expect(state.products).toEqual(products);
+    });
+  });
+
+  describe('setProduct', () => {
+    const initialState = {
+      product: null,
+    };
+
+    const product = products[0];
+
+    const state = reducer(initialState, setProduct(product));
+
+    expect(state.product.id).toBe(1);
+    expect(state.product.title).toBe('크리넥스 KF-AD 소형 마스크 팝니다.');
+  });
+});
+
+describe('actions', () => {
+  let store;
+
+  describe('loadInitProducts', () => {
+    beforeEach(() => {
+      store = mockStore({});
+    });
+
+    it('runs setProducts', async () => {
+      await store.dispatch(loadInitProducts());
+
+      const actions = store.getActions();
+
+      expect(actions[0]).toEqual(setProducts([]));
+    });
+  });
+
+  describe('loadProduct', () => {
+    beforeEach(() => {
+      store = mockStore({});
+    });
+
+    it('dispatchs setProduct', async () => {
+      await store.dispatch(loadProduct({ productId: 1 }));
+
+      const actions = store.getActions();
+
+      expect(actions[0]).toEqual(setProduct({}));
+    });
+  });
+});
